+++ src/seabolt/src/bolt/config-private.h
@@ -23,16 +23,16 @@
 
 struct BoltTrust {
     char* certs;
-	int certs_len;
-    int skip_verify;
-    int skip_verify_hostname;
+	uint64_t certs_len;
+    int32_t skip_verify;
+    int32_t skip_verify_hostname;
 };
 
 struct BoltSocketOptions {
-    int connect_timeout;
-    int recv_timeout;
-    int send_timeout;
-    int keep_alive;
+	int32_t connect_timeout;
+	int32_t recv_timeout;
+	int32_t send_timeout;
+	int32_t keep_alive;
 };
 
 struct BoltConfig {
@@ -43,9 +43,9 @@ struct BoltConfig {
     struct BoltValue* routing_context;
     struct BoltAddressResolver* address_resolver;
     struct BoltLog* log;
-    int max_pool_size;
-    int max_connection_life_time;
-    int max_connection_acquisition_time;
+	int32_t max_pool_size;
+	int32_t max_connection_life_time;
+	int32_t max_connection_acquisition_time;
     struct BoltSocketOptions* socket_options;
 };
